removed rsk-wallet class and included rsk functionality in wallet.
