+++ lib/api_client.js
@@ -1282,114 +1282,63 @@ APIClient.prototype.initWallet = function(options, cb) {
                 throw new Error("Backup key returned from server didn't match our own copy");
             }
         }
-        var backupPublicKey = bitcoin.HDNode.fromBase58(result.backup_public_key[0], self.network);
-        var blocktrailPublicKeys = _.mapValues(result.blocktrail_public_keys, function(blocktrailPublicKey) {
-            return bitcoin.HDNode.fromBase58(blocktrailPublicKey[0], self.network);
-        });
-        var primaryPublicKeys = _.mapValues(result.primary_public_keys, function(primaryPublicKey) {
+
+        var backupPublicKey,blocktrailPublicKeys,primaryPublicKeys, wallet;
+
+        primaryPublicKeys = _.mapValues(result.primary_public_keys, function(primaryPublicKey) {
             return bitcoin.HDNode.fromBase58(primaryPublicKey[0], self.network);
         });
-
-        // initialize wallet
-        var wallet = new Wallet(
-            self,
-            identifier,
-            options.walletVersion,
-            result.primary_mnemonic,
-            result.encrypted_primary_seed,
-            result.encrypted_secret,
-            primaryPublicKeys,
-            backupPublicKey,
-            blocktrailPublicKeys,
-            keyIndex,
-            result.segwit || 0,
-            self.testnet,
-            self.regtest,
-            result.checksum,
-            result.upgrade_key_index,
-            options.useCashAddress,
-            options.bypassNewAddressCheck
-        );
-
-        wallet.recoverySecret = result.recovery_secret;
-
-        if (!options.readOnly) {
-            return wallet.unlock(options).then(function() {
-                return wallet;
+        
+        if (self.chain === "RSK") 
+        {
+            wallet = new Wallet(
+                self,
+                identifier,
+                options.walletVersion,
+                result.primary_mnemonic,
+                result.encrypted_primary_seed,
+                result.encrypted_secret,
+                primaryPublicKeys,
+                backupPublicKey,
+                blocktrailPublicKeys,
+                keyIndex,
+                result.segwit || 0,
+                self.testnet,
+                self.regtest,
+                result.checksum,
+                result.upgrade_key_index,
+                options.useCashAddress,
+                options.bypassNewAddressCheck
+            );
+        }
+        else
+        {
+            backupPublicKey = bitcoin.HDNode.fromBase58(result.backup_public_key[0], self.network);
+            blocktrailPublicKeys = _.mapValues(result.blocktrail_public_keys, function(blocktrailPublicKey) {
+                return bitcoin.HDNode.fromBase58(blocktrailPublicKey[0], self.network);
             });
-        } else {
-            return wallet;
+
+             // initialize wallet
+            wallet = new Wallet(
+                self,
+                identifier,
+                options.walletVersion,
+                result.primary_mnemonic,
+                result.encrypted_primary_seed,
+                result.encrypted_secret,
+                primaryPublicKeys,
+                backupPublicKey,
+                blocktrailPublicKeys,
+                keyIndex,
+                result.segwit || 0,
+                self.testnet,
+                self.regtest,
+                result.checksum,
+                result.upgrade_key_index,
+                options.useCashAddress,
+                options.bypassNewAddressCheck
+            );
         }
-    }));
-
-    return deferred.promise;
-};
-
-/**
- * initialize an existing RSK wallet
- *
- * Either takes two argument:
- * @param options       object      {}
- * @param [cb]          function    callback(err, wallet, primaryMnemonic, backupMnemonic, blocktrailPubKeys)
- *
- * @returns {q.Promise}
- */
-APIClient.prototype.initRskWallet = function(options, cb) {
-    var self = this;
-
-    if (typeof options !== "object") {
-        // get the old-style arguments
-        options = {
-            identifier: arguments[0],
-            passphrase: arguments[1]
-        };
-
-        cb = arguments[2];
-    }
-
-    if (options.check_backup_key) {
-        if (typeof options.check_backup_key !== "string") {
-            throw new Error("Invalid input, must provide the backup key as a string (the xpub)");
-        }
-    }
-
-    var deferred = q.defer();
-    deferred.promise.spreadNodeify(cb);
-
-    var identifier = options.identifier;
-
-    if (!identifier) {
-        deferred.reject(new blocktrail.WalletInitError("Identifier is required"));
-        return deferred.promise;
-    }
-
-    deferred.resolve(self.blocktrailClient.get("/wallet/" + identifier, null, true).then(function(result) {
-        var keyIndex = options.keyIndex || result.key_index;
-
-        options.walletVersion = result.wallet_version;
-
-        var primaryPublicKeys = _.mapValues(result.primary_public_keys, function(primaryPublicKey) {
-            return bitcoin.HDNode.fromBase58(primaryPublicKey[0], self.network);
-        });
-
-        // initialize wallet
-        var wallet = new RskWallet(
-            self,
-            identifier,
-            options.walletVersion,
-            result.primary_mnemonic,
-            result.encrypted_primary_seed,
-            result.encrypted_secret,
-            primaryPublicKeys,
-            keyIndex,
-            result.segwit || 0,
-            self.testnet,
-            self.regtest,
-            result.checksum,
-            result.upgrade_key_index,
-            options.useCashAddress,
-            options.bypassNewAddressCheck
-        );
 
         wallet.recoverySecret = result.recovery_secret;
 
@@ -2283,7 +2232,7 @@ APIClient.prototype.sendTransaction = function(identifier, txHex, paths, checkFe
     );
 };
 
-APIClient.prototype.sendRskTransaction = function(identifier, txHex, paths, checkFee, twoFactorToken, prioboost, options, cb) {
+APIClient.prototype.sendRskTransaction = function(identifier, txHex, checkFee, twoFactorToken, prioboost, options, cb) {
     var self = this;
 
     if (typeof twoFactorToken === "function") {
@@ -2299,7 +2248,7 @@ APIClient.prototype.sendRskTransaction = function(identifier, txHex, paths, chec
     }
 
     var data = {
-        paths: paths,
+        // paths: paths,
         two_factor_token: twoFactorToken
     };
     if (typeof txHex === "string") {
@@ -2314,20 +2263,6 @@ APIClient.prototype.sendRskTransaction = function(identifier, txHex, paths, chec
         check_fee: checkFee ? 1 : 0,
         prioboost: prioboost ? 1 : 0
     };
-
-    var bip70 = false;
-    if (options.bip70PaymentUrl) {
-        bip70 = true;
-        postOptions.bip70PaymentUrl = options.bip70PaymentUrl;
-
-        if (options.bip70MerchantData && options.bip70MerchantData instanceof Uint8Array) {
-            // Encode merchant data to base64
-            var decoder = new TextDecoder('utf8');
-            var bip70MerchantData = btoa(decoder.decode(options.bip70MerchantData));
-
-            postOptions.bip70MerchantData = bip70MerchantData;
-        }
-    }
 
     return self.blocktrailClient.post(
         "/wallet/" + identifier + "/send",
