+++ lib/wallet.js
@@ -11,6 +11,11 @@ var EncryptionMnemonic = require('./encryption_mnemonic');
 var SizeEstimation = require('./size_estimation');
 var bip39 = require('bip39');
 
+// Libraries for rootstock, ethereum key management.
+var ethereumKey = require('ethereumjs-wallet/hdkey')
+var ethutil = require('ethereumjs-util')
+var ethtx = require('ethereumjs-tx')
+
 var SignMode = {
     SIGN: "sign",
     DONT_SIGN: "dont_sign"
@@ -90,9 +95,12 @@ var Wallet = function(
         }
     }
 
-    assert(backupPublicKey instanceof bitcoin.HDNode);
     assert(_.every(primaryPublicKeys, function(primaryPublicKey) { return primaryPublicKey instanceof bitcoin.HDNode; }));
-    assert(_.every(blocktrailPublicKeys, function(blocktrailPublicKey) { return blocktrailPublicKey instanceof bitcoin.HDNode; }));
+
+    if (self.sdk.chain != "RSK") {
+        assert(backupPublicKey instanceof bitcoin.HDNode);
+        assert(_.every(blocktrailPublicKeys, function(blocktrailPublicKey) { return blocktrailPublicKey instanceof bitcoin.HDNode; }));
+    }
 
     // v1
     self.primaryMnemonic = primaryMnemonic;
@@ -148,6 +156,7 @@ Wallet.PAY_PROGRESS_DONE = 100;
 Wallet.CHAIN_BTC_DEFAULT = 0;
 Wallet.CHAIN_BTC_SEGWIT = 2;
 Wallet.CHAIN_BCC_DEFAULT = 1;
+Wallet.CHAIN_RSK_DEFAULT = 5;
 
 Wallet.FEE_STRATEGY_FORCE_FEE = blocktrail.FEE_STRATEGY_FORCE_FEE;
 Wallet.FEE_STRATEGY_BASE_FEE = blocktrail.FEE_STRATEGY_BASE_FEE;
@@ -693,6 +702,80 @@ Wallet.prototype.upgradeKeyIndex = function(keyIndex, cb) {
     );
 
     return deferred.promise;
+};
+
+/**
+ * generate a new derived private key and return the rootstock address for it
+ *
+ * @param [chainIdx] int
+ * @param [cb]  function        callback(err, address)
+ * @returns {q.Promise}
+ */
+Wallet.prototype.getRskAddress = function(chainIdx, cb) {
+    var self = this;
+
+    // chainIdx is optional
+    if (typeof chainIdx === "function") {
+        cb = chainIdx;
+        chainIdx = null;
+    }
+
+    var deferred = q.defer();
+    deferred.promise.spreadNodeify(cb);
+
+    // Only enter if it's not an integer
+    if (chainIdx !== parseInt(chainIdx, 10)) {
+        // deal with undefined or null, assume defaults
+        if (typeof chainIdx === "undefined" || chainIdx === null) {
+            chainIdx = Wallet.CHAIN_RSK_DEFAULT;
+        } else {
+            // was a variable but not integer
+            deferred.reject(new Error("Invalid chain index"));
+            return deferred.promise;
+        }
+    }
+
+    // Get Rootstock address from server.
+    deferred.resolve(self.sdk.getRskNewDerivation(self.identifier, "M/" + self.keyIndex + "'/" + chainIdx)
+        .then(function(newDerivation) {
+
+            var path = newDerivation.path.replace("\'","").replace('9999',"9999'");
+            var addressFromServer = newDerivation.address;
+
+            // Verify Rootstock address client-side.
+            var verifyAddress = self.getRskAddressByPath(path);
+            
+            if (verifyAddress !== addressFromServer) {
+                throw new blocktrail.WalletAddressError("Failed to verify address [" + addressFromServer + "] !== [" + verifyAddress + "]");
+            }
+
+            return [verifyAddress, path];
+        }));
+    
+    return deferred.promise;
+};
+
+/**
+ * get rootstock address for specified path
+ *
+ * @param path
+ * @returns string
+ */
+Wallet.prototype.getRskAddressByPath = function(path) {
+    var self = this;
+
+    // Get the derived primary public key.
+    var derivedPrimaryPublicKey = self.getPrimaryPublicKey(path);
+    
+    // Encode the derived key to base58.
+    var encodedPublicKey = derivedPrimaryPublicKey.toBase58();
+
+    // Instantiate an HD ethereum key using the extended public key.
+    var derivedEthereumKey = ethereumKey.fromExtendedKey(encodedPublicKey);
+
+    // Use the public key to encode an ethereum address (without address checksum).
+    var address = derivedEthereumKey.getWallet().getAddressString();
+    return address;
 };
 
 /**
@@ -1884,6 +1967,252 @@ Wallet.deriveByPath = function(hdKey, path, keyPath) {
     } else {
         return newKey;
     }
+};
+
+Wallet.prototype.rskPay = function(pay, changeAddress, allowZeroConf, randomizeChangeIdx, feeStrategy, twoFactorToken, options, cb) {
+
+    /* jshint -W071 */
+    var self = this;
+
+    if (typeof changeAddress === "function") {
+        cb = changeAddress;
+        changeAddress = null;
+    } else if (typeof allowZeroConf === "function") {
+        cb = allowZeroConf;
+        allowZeroConf = false;
+    } else if (typeof randomizeChangeIdx === "function") {
+        cb = randomizeChangeIdx;
+        randomizeChangeIdx = true;
+    } else if (typeof feeStrategy === "function") {
+        cb = feeStrategy;
+        feeStrategy = null;
+    } else if (typeof twoFactorToken === "function") {
+        cb = twoFactorToken;
+        twoFactorToken = null;
+    } else if (typeof options === "function") {
+        cb = options;
+        options = {};
+    }
+
+    randomizeChangeIdx = typeof randomizeChangeIdx !== "undefined" ? randomizeChangeIdx : true;
+    feeStrategy = feeStrategy || Wallet.FEE_STRATEGY_OPTIMAL;
+    options = options || {};
+    var checkFee = typeof options.checkFee !== "undefined" ? options.checkFee : true;
+
+    var deferred = q.defer();
+    deferred.promise.nodeify(cb);
+
+    if (self.locked) {
+        deferred.reject(new blocktrail.WalletLockedError("Wallet needs to be unlocked to send coins"));
+        return deferred.promise;
+    }
+
+    q.nextTick(function() {
+        
+        self.buildRskTransaction(pay, changeAddress, allowZeroConf, randomizeChangeIdx, feeStrategy, options).then(
+            function(r) { return r; },
+            function(e) { deferred.reject(e); },
+        )
+            .spread(
+            function(tx) {
+
+                // var data = {
+                //     signed_transaction: tx.toHex(),
+                //     base_transaction: tx.__toBuffer(null, null, false).toString('hex')
+                // };
+
+                console.log(tx)
+
+                return self.sendRskTransaction(tx, checkFee, twoFactorToken, options.prioboost, options)
+                    .then(function(result) {
+                        
+
+                        if (!result || !result['complete'] || result['complete'] === 'false') {
+                            deferred.reject(new Error("Failed to send transaction."));
+                        } else {
+                            return result['txid'];
+                        }
+                    });
+            },
+            function(e) {
+                throw e;
+            }
+        )
+            .then(
+            function(r) { deferred.resolve(r); },
+            function(e) { deferred.reject(e); }
+        )
+        ;
+    });
+
+    return deferred.promise;
+};
+
+Wallet.prototype.buildRskTransaction = function(pay, changeAddress, allowZeroConf, randomizeChangeIdx, feeStrategy, options, cb) {
+    /* jshint -W071 */
+    var self = this;
+
+    if (typeof changeAddress === "function") {
+        cb = changeAddress;
+        changeAddress = null;
+    } else if (typeof allowZeroConf === "function") {
+        cb = allowZeroConf;
+        allowZeroConf = false;
+    } else if (typeof randomizeChangeIdx === "function") {
+        cb = randomizeChangeIdx;
+        randomizeChangeIdx = true;
+    } else if (typeof feeStrategy === "function") {
+        cb = feeStrategy;
+        feeStrategy = null;
+    } else if (typeof options === "function") {
+        cb = options;
+        options = {};
+    }
+
+    randomizeChangeIdx = typeof randomizeChangeIdx !== "undefined" ? randomizeChangeIdx : true;
+    feeStrategy = feeStrategy || Wallet.FEE_STRATEGY_OPTIMAL;
+    options = options || {};
+
+    
+    
+    var deferred = q.defer();
+
+    // @TODO - get chainIdx and childIdx at init method
+    var path = "M/" + self.keyIndex + "'/" + 5 + "/" + 0;
+    console.log(path);
+    var accountAddress = self.getAddressByPath(path);
+
+    q.nextTick(function() {
+        deferred.resolve(
+            self.getRskAccount(accountAddress).spread(function(confirmed_balance,address,path,nonce){
+                var txb, serializedTx;
+                var deferred = q.defer();
+                async.waterfall([
+                    /**
+                     * init transaction builder
+                     *
+                     * @param cb
+                     */
+                    function(cb) {
+                        // blank ethereum transaction
+                        // @TODO - other chain ids - testnet,regtest for rsk?
+                        // chainId 31 is for mainnet rsk
+                        txb = new ethtx(null,31);
+
+                        cb();
+                    },
+                    /**
+                     * add transaction params
+                     *
+                     * @param cb
+                     */
+                    function(cb) {
+                        txb.nonce = 0;
+                        txb.gasPrice = 10;
+                        txb.gasLimit = 100;
+                        txb.value = 0;
+                        txb.data = 0;
+                        txb.to = "0x0000000000000000000000000000000000000000"
+
+                        cb();
+                    },
+                    /**
+                     * estimate fee to verify that the API is not providing us wrong data
+                     *
+                     * @param cb
+                     */
+                    function(cb) {
+                        // use web3.eth.gasPrice to get gas price for given network
+                        cb();
+                    },
+                    /**
+                     * sign transaction
+                     *
+                     * @param cb
+                     */
+                    function(cb) {
+                        var i, privKey
+
+                        path = path.replace("M","m");
+
+                        
+                        console.log(self.primaryPrivateKey);
+                        // todo: regenerate scripts for path and compare for utxo (paranoid mode)
+                        if (self.primaryPrivateKey) {
+                            privKey = self.deriveByPath(self.primaryPrivateKey, path, "m");
+                        }
+                        else {
+                            throw new Error("No master privateKey present");
+                        }
+                        txb.sign(ethutil.toBuffer("0x"+privKey.keyPair.d.toBuffer().toString('hex')));
+                        serializedTx = txb.serialize();
+
+                        cb();
+                    },
+                ], function(err) {
+                    if (err) {
+                        deferred.reject(new blocktrail.WalletSendError(err));
+                        return;
+                    }
+
+                    deferred.resolve([serializedTx]);
+                });
+                
+                return deferred.promise;
+            })
+        )
+    });
+
+    return deferred.promise;
+};
+
+Wallet.prototype.sendRskTransaction = function(tx, checkFee, twoFactorToken, prioboost, options, cb) {
+    var self = this;
+
+    if (typeof twoFactorToken === "function") {
+        cb = twoFactorToken;
+        twoFactorToken = null;
+        prioboost = false;
+    } else if (typeof prioboost === "function") {
+        cb = prioboost;
+        prioboost = false;
+    } else if (typeof options === "function") {
+        cb = options;
+        options = {};
+    }
+
+    var deferred = q.defer();
+    deferred.promise.nodeify(cb);
+
+    self.sdk.sendRskTransaction(self.identifier, tx.toString('hex'), checkFee, twoFactorToken, prioboost, options)
+        .then(
+            function(result) {
+                deferred.resolve(result);
+            },
+            function(e) {
+                if (e.requires_2fa) {
+                    deferred.reject(new blocktrail.WalletMissing2FAError());
+                } else if (e.message.match(/Invalid two_factor_token/)) {
+                    deferred.reject(new blocktrail.WalletInvalid2FAError());
+                } else {
+                    deferred.reject(e);
+                }
+            }
+        )
+    ;
+
+    return deferred.promise;
+};
+
+Wallet.prototype.getRskAccount = function(address, options, cb) {
+    var self = this;
+
+    if (typeof options === "function") {
+        cb = options;
+        options = {};
+    }
+
+    return self.sdk.getRskAccount(self.identifier, address, options, cb);
 };
 
 module.exports = Wallet;
